- add valid date filter - fix deletion of entities through admin
git-svn-id: http://projects.conceptive.be/camelot/svn/trunk@1014 8325946a-aa55-0410-8760-f0bc8c33efaa

+++ camelot/view/elixir_admin.py
@@ -575,16 +575,15 @@ class EntityAdmin(ObjectAdmin):
     def delete(self, entity_instance):
         """Delete an entity instance"""
         from sqlalchemy.orm.session import Session
-        if entity_instance.id:
+        if hasattr(entity_instance, 'id') and entity_instance.id:
                 pk = entity_instance.id
                 # save the state before the update
                 from camelot.model.memento import BeforeDelete
                 from camelot.model.authentication import getCurrentAuthentication
-                history = BeforeDelete( model = unicode( self.admin.entity.__name__ ),
+                history = BeforeDelete( model = unicode( self.entity.__name__ ),
                                        primary_key = pk,
                                        previous_attributes = {},
                                        authentication = getCurrentAuthentication() )
-                self.logger.debug( 'delete the object' )
                 entity_instance.delete()
                 Session.object_session( entity_instance ).flush( [entity_instance] )
                 Session.object_session( history ).flush( [history] )
